Flatten raw assets to top level.
Building the site now outputs to a dist folder. Also removed the
uploading to gh-pages step for now as this work is intended to set up
hosting on netlify first.  If netlify ends up sucking or going out of
buisness we can go back to using http://ghub.io/gh-pages